Add data layer to location module

+++ app/src/main/kotlin/cz/cvut/zan/zavadmak/weatherapp/location/data/remote/model/LocationDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class LocationDto(
     @SerialName("place_id")
-    val placeId: Int,
+    val placeId: Long,
     val lat: Double,
     val lon: Double,
     val name: String,
