+++ app/src/main/kotlin/cz/cvut/zan/zavadmak/weatherapp/location/domain/usecase/RemoveLocationUseCase.kt
@@ -2,6 +2,6 @@ package cz.cvut.zan.zavadmak.weatherapp.location.domain.usecase
 
 interface RemoveLocationUseCase {
 
-    suspend fun execute(locationId: Int)
+    suspend fun execute(locationId: Long)
 
 }
