+++ app/src/main/kotlin/cz/cvut/zan/zavadmak/weatherapp/location/domain/usecase/RemoveLocationUseCaseImpl.kt
@@ -1,7 +1,7 @@
 package cz.cvut.zan.zavadmak.weatherapp.location.domain.usecase
 
 class RemoveLocationUseCaseImpl : RemoveLocationUseCase {
-    override suspend fun execute(locationId: Int) {
+    override suspend fun execute(locationId: Long) {
 
     }
 }
