+++ app/src/main/kotlin/cz/cvut/zan/zavadmak/weatherapp/weather/presentation/navigation/WeatherScreens.kt
@@ -5,9 +5,9 @@ import kotlinx.serialization.Serializable
 sealed interface WeatherScreens {
 
     @Serializable
-    data class CurrentWeather(val locationId: Int) : WeatherScreens
+    data class CurrentWeather(val locationId: Long) : WeatherScreens
 
     @Serializable
-    data class Forecast(val locationId: Int) : WeatherScreens
+    data class Forecast(val locationId: Long) : WeatherScreens
 
 }
